storage: Split pkg cloud into cloud interface and cloud impl
Previously, both the ExternalStorage interface and all the specialized
implementations were in `pkg/storage/cloud`.

This PR moves the interface and factory method signatures to
`pkg/storage/cloud` and all the implementations to
`pkg/storage/cloudimpl`.  The motivation behind this is to ensure that
other packages such `pkg/sql` and `pkg/kv` depend on just the
interface/factory signatures (not the concrete impl) to prevent
dependency cycles.

More concreteley, we need to plumb an InternalExecutor and kv.DB for our
new user scoped file-table storage. This was previously not possible
because of a `pkg/sql -> pkg/storage/cloud ->pkg/sql` cycle. Now, since
the executor will be used in `cloudimpl` this cycle will not occur.

Release note: None
