+++ pkg/ccl/importccl/read_import_base.go
@@ -32,6 +32,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 	"github.com/cockroachdb/cockroach/pkg/storage/cloud"
+	"github.com/cockroachdb/cockroach/pkg/storage/cloudimpl"
 	"github.com/cockroachdb/cockroach/pkg/util/ctxgroup"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/tracing"
@@ -135,7 +136,7 @@ func readInputFiles(
 
 	// Attempt to fetch total number of bytes for all files.
 	for id, dataFile := range dataFiles {
-		conf, err := cloud.ExternalStorageConfFromURI(dataFile)
+		conf, err := cloudimpl.ExternalStorageConfFromURI(dataFile)
 		if err != nil {
 			return err
 		}
@@ -160,7 +161,7 @@ func readInputFiles(
 		default:
 		}
 		if err := func() error {
-			conf, err := cloud.ExternalStorageConfFromURI(dataFile)
+			conf, err := cloudimpl.ExternalStorageConfFromURI(dataFile)
 			if err != nil {
 				return err
 			}
@@ -213,7 +214,7 @@ func readInputFiles(
 					if err != nil {
 						return err
 					}
-					conf, err := cloud.ExternalStorageConfFromURI(rejFn)
+					conf, err := cloudimpl.ExternalStorageConfFromURI(rejFn)
 					if err != nil {
 						return err
 					}
