+++ pkg/testutils/lint/lint_test.go
@@ -438,7 +438,7 @@ func TestLint(t *testing.T) {
 					":!acceptance",
 					":!ccl/acceptanceccl/backup_test.go",
 					":!ccl/backupccl/backup_cloud_test.go",
-					":!storage/cloud",
+					":!storage/cloudimpl",
 					":!ccl/workloadccl/fixture_test.go",
 					":!cmd",
 					":!nightly",
